requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
SQLAlchemy==2.0.21
cryptography==41.0.7
Werkzeug==2.3.7
bcrypt==4.0.1
python-dotenv==1.0.0
joblib==1.3.2
google-generativeai==0.3.0
requests==2.31.0
simplejson==3.19.1